requirements.txt
aiogram>=3.13
aiohttp>=3.11
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
aiosqlite>=0.20
python-dotenv>=1.0
aiofiles>=24.1
anthropic>=0.42
